keep count in field of type long instead of int
git-svn-id: https://src.springframework.org/svn/spring-maintenance/trunk@9428 fd5a2b45-1f63-4059-99e9-3c7cb7fd75c8

+++ src/org/springframework/aop/interceptor/DebugInterceptor.java
@@ -34,7 +34,7 @@ import org.aopalliance.intercept.MethodInvocation;
  */
 public class DebugInterceptor extends SimpleTraceInterceptor {
 
-	private volatile int count;
+	private volatile long count;
 
 
 	/**
@@ -69,7 +69,7 @@ public class DebugInterceptor extends SimpleTraceInterceptor {
 	/**
 	 * Return the number of times this interceptor has been invoked.
 	 */
-	public int getCount() {
+	public long getCount() {
 		return this.count;
 	}
